requirements.txt
python-telegram-bot==12.7
pickle
matplotlib
pycodestyle